requirements.txt
langchain
langchain-community
langchain-core
mistralai
streamlit
python-dotenv